+++ services/user/Homepage/ui/src/apps/chainmail/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./compose-dialog";
-export * from "./empty-box";
+export * from "./empty-states";
 export * from "./mail-list";
 export * from "./mailbox";
 export * from "./message-detail";
